Refactor DataContainer I/O
